Final 2017 update used at Christmas
This is the final version of the WS2812 controller used with SmartThings at Christmas 2017.  These controllers would have problems after a period of time and this has never been diagnosed. There is a possibility that the problem might be due to losing WiFi connection so this will need to be investigated next year.